+++ core/assets/characters/images/SlimeSS.png
