+++ core/src/com/icnhdevelopment/wotn/Game.java
@@ -3,24 +3,16 @@ package com.icnhdevelopment.wotn;
 import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.*;
-import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.math.Vector2;
 import com.icnhdevelopment.wotn.gui.*;
-import com.icnhdevelopment.wotn.gui.Container;
-import com.icnhdevelopment.wotn.gui.Label;
 import com.icnhdevelopment.wotn.handlers.CInputProcessor;
 import com.icnhdevelopment.wotn.handlers.ColorCodes;
-
-import java.util.Date;
-import java.util.Timer;
-import java.util.TimerTask;
 
 public class Game extends ApplicationAdapter {
 
 	SpriteBatch batch;
 	CInputProcessor inputProcessor;
-	static Container main;
+	static Menu currentMenu;
 
 	static int WIDTH, HEIGHT;
 
@@ -42,56 +34,8 @@ public class Game extends ApplicationAdapter {
 		Gdx.input.setInputProcessor(inputProcessor);
 		WIDTH = Gdx.graphics.getWidth();
 		HEIGHT = Gdx.graphics.getHeight();
-		main = new Container();
-		Label l = new Label(main, new Vector2(0, Game.HEIGHT()-44), new Vector2(Game.WIDTH(), 40), "The Wizard of the North");
-		l.setFontSize(40);
-		l.setColor(Color.WHITE);
-		l.setBorderColor(Color.RED);
-		l.sethAlignment(Alignment.CENTER);
-		l.setBackColor(Color.BLUE);
-		l.createFont(false);
-		Button b = new Button(main, new Vector2(0, 0), new Vector2(4, 4), "Exit") {
-			@Override
-			public void Click() {
-				Game.EXIT();
-			}
-		};
-		b.setFontSize(30);
-		b.setColor(Color.BLACK);
-		b.setBorderColor(Color.GREEN);
-		b.setBackColor(Color.YELLOW);
-		//b.setvAlignment(Alignment.MIDDLE);
-		//b.sethAlignment(Alignment.CENTER);
-		b.createFont(true);
-		main.buttons.add(b);
-		ImageLabel il = new ImageLabel(main, new Vector2(300, 300), new Vector2(256, 256), new Texture(Gdx.files.internal("badlogic.jpg")));
-		il.setImageAlignment(Alignment.STRETCHED);
-		il.setBackColor(Color.RED);
-		il.setBackColorOpacity(.2f);
-		b = new Button(il, new Vector2(10, 10), new Vector2(2, 2), "Shrink :D") {
-			@Override
-			public void Click() {
-				new Timer().schedule(new TimerTask() {
-					@Override
-					public void run() {
-						if (main.getSize().x > 800) {
-							main.getAbsoluteParent().resize(main.getSize().x-14, main.getSize().y);
-						}
-						if (main.getSize().y > 450) {
-							main.getAbsoluteParent().resize(main.getSize().x, main.getSize().y-8);
-						}
-						if (main.getSize().x <= 800 && main.getSize().y <= 450) {
-							this.cancel();
-						}
-					}
-				}, new Date(), 20);
-			}
-		};
-		b.setBackColor(Color.ORANGE);
-		b.sethAlignment(Alignment.CENTER);
-		b.setvAlignment(Alignment.MIDDLE);
-		b.createFont(true);
-		main.buttons.add(b);
+		currentMenu = new Menu();
+		currentMenu.init();
 	}
 
 	@Override
@@ -100,14 +44,9 @@ public class Game extends ApplicationAdapter {
 		Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT | GL20.GL_DEPTH_BUFFER_BIT);
 		Gdx.gl.glEnable(GL20.GL_BLEND);
 		Gdx.gl.glBlendFunc(GL20.GL_SRC_ALPHA, GL20.GL_ONE_MINUS_SRC_ALPHA);
-		main.updateChildren(inputProcessor);
-		batch.setProjectionMatrix(main.getRenderCam().combined);
-		batch.begin();
-		main.renderBackground(batch);
-		batch.end();
-		batch.begin();
-		main.render(batch);
-		batch.end();
+		currentMenu.update(inputProcessor);
+		batch.setProjectionMatrix(currentMenu.mainContainer.getRenderCam().combined);
+		currentMenu.render(batch);
 		inputProcessor.update();
 		Gdx.gl.glDisable(GL20.GL_BLEND);
 	}
